server/repo: move rev_list_raw to LocalRepo
This really isn't ideal, it only exists for the server to
stream out the data to the rev-list command. We really
should refactor this instead, so that the existing rev-list
command is provided on top of a pre-parsed rev_list() API,
which also implies internal API changes.

For now, just move it. We can do the refactoring later.
